[Pal/Linux-SGX] Allow returning to different stack in _restore_sgx_context()
Previously, Linux-SGX PAL _restore_sgx_context(uc) assumed that the
context uc is allocated on the same normal stack and that this stack
is contiguous; in particular, see _restore_sgx_context(uc).

However, a dedicated signal stack introduced in the next commits is
non-contiguous with normal stack. Thus, the context and the returning
stack frame may be allocated on a different stack. This commit modifies
_restore_sgx_context() to allow uc from a different stack, and to return
to different stack via `jmp *tmp_rip` instead of `ret`.