[Pal/Linux-SGX] Make pal-sgx-sign code importable
This moves the Python implementation of pal-sgx-sign into
pal_sgx_sign.py and replaces the cmdline tool with a tiny wrapper. This
allows other code to import it as a module and thereby re-use its code.

Designing an API and then putting it into a proper Python package is left
for the future.