[Pal/Linux-SGX] Introduce sgx_get_quote() OCALL
New sgx_get_quote() OCALL allows to retrieve the SGX Quote from the
Quoting Enclave. The retrieved quote is based on this enclave's
SGX report and corresponds to this enclave. The caller of this OCALL
may forward the quote to the remote user for remote attestation.
A follow-up commit will introduce the actual callers.