[Pal/Linux-SGX] tools: add RA-TLS utility libraries
RA-TLS integrates Intel SGX remote attestation into the TLS connection
setup. Conceptually, it extends the standard X.509 certificate with
SGX-related information. The additional information allows the receiver
(verifier) of the certificate to verify that it is indeed communicating
with an SGX enclave (attester). RA-TLS is shipped as three libraries:

- ra_tls_attest.so: creates self-signed RA-TLS X.509 certificate with
  SGX quote embedded; typically linked into server apps.
- ra_tls_verify_epid.so: verifies RA-TLS certificate by sending SGX
  quote to IAS and verifying attestation report from IAS; typically
  linked into client apps.
- ra_tls_verify_dcap.so: verifies RA-TLS certificate by providing SGX
  quote v3 to the libsgx_dcap_quoteverify library and relying on its
  assessment of the quote; typically linked into client apps.

This commit also adds an RA-TLS example using simple mbedTLS server
and client. The server generates the RA-TLS certificate, and the client
connects to the server, verifies this certificate, and performs a
dummy HTTP request. This example is added to Jenkins.