[Pal] Enforce that executable is always located at a predefined address
Previously, the PAL layer would put an executable at a first unoccupied
address range if the executable was position-independent (PIE). This
could lead to the same executable being located at different addresses
across forks (Graphene correctly checkpoints shared libraries but
has separate handling for the executable). This is a rare scenario
since the PAL initialization code is typically deterministic across
forks. However, rarely there would be small difference in allocations
which would lead to different base addresses for executable segments
in parent and child processes, and segfaults and data corruptions
happened (this was the case for Nginx on Ubuntu 18.04 with change in
the manifest file processing). This commit simply forces executables
to always be loaded at a predefined address (currently 0x00400000).