[Pal] Force PIE executable to be located at address 0x555555554000
Previously, we forced PIE executables to be located at address
0x00400000 (4MB). However, there is a bug in the ELF relocation code
in LibOS that leads to double-relocation. To circumvent this bug,
relocation code checks if the offset was already relocated. But if
the offset itself exceeds the base address of PIE executable (i.e.,
exceeds 4MB), then the offset is not relocated and segfaults follow.
This commit changes base address from 0x00400000 to 0x555555554000,
similar to what Linux does.