[Pal/Linux-SGX] sgx-gdb: Fix single-stepping by jumping to ERESUME
Commit 3178e26 "Add sgx.print_stats manifest option to display SGX
stats" broke the SGX-GDB ptrace wrapper by adding a new instruction
to the Asynchronous Exit Pointer assembly. This new instruction
confused the GDB wrapper when it tried to single-step using ptrace's
SINGLESTEP command: it assumed that the only instruction in AEP is
ERESUME. This commit adds new function entry to the AEP assembly so
that the GDB wrapper moves RIP to ERESUME and continues single-
stepping from the correct position. This commit also fixes small
bug with thread_stepping by converting it to a proper bit vector.