[LibOS] Do not try to send signal 0 in kill()/tkill()
Previously, Graphene didn't handle special case of kill()/tkill()
with signal equal to zero. This signal is used to check for existence
of processes and threads. However, Graphene tried to send/append it
which led to segfaults. This commit correctly handles signal equal
to zero and improves LibOS tests to verify this fix.