[LibOS] Hide arguments from host OS and send them in checkpoint
Previously, Graphene sent arguments to a child process (spawned via
execve() syscall) in the clear. This was insecure. This commit fixes
it by always spawning a new child process without arguments and then
sending the arguments in the checkpoint (which is securely encrypted
on Linux-SGX PAL). The new logic is the same as with environ.