[Pal/Linux-SGX] Preserve wrap key for protected files across forks
Trusted child enclaves must inherit the provisioned master (wrap) key
for protected files from the parent enclave on fork/clone. This commit
adds this functionality and modifies the RA-TLS secret provisioning
PF client to test this (including in Jenkins).

Kudos to Li Xun for reporting this bug and proposing a solution.