[Pal/Linux-SGX] Allocate min required memory for file's filename
Previously, Linux-SGX PAL allocated 4KB of memory for URI (filename)
of any trusted/allowed file or directory. This led to "out of PAL
memory" errors in Graphene if there are 100,000s of trusted files,
as is the case for complex workloads in GSC/Docker. This commit
removes the static 4KB buffer in file's metadata and instead
allocates the minimum required buffer. This commit also cleans up
`register_trusted_file()` code a bit.