[LibOS] Rework sending memory content in checkpoints
Previously memory content was sent inside a checkpoint and counted
towards the checkpoint size in the child process, but not in the parent
process. The child allocated the checkpoint at the same address as in
the parent, but it had a different size, so it could colide with another
memory mapping. This commmit fixes it by making checkpointing code
receive the memory directly to destination addresses (which additionally
saves an unnecessary copy).