[LibOS] Silence clang warning about GNU alignof usage
Clang complains that using alignof on an expression is a GNU
extension. We use it in constructions such as (see shim_futex.c):

    if (!IS_ALIGNED_PTR(ptr, alignof(*ptr)))

This could be rewritten to use the actual type
(IS_ALIGNED_PTR(ptr, alignof(uint32_t))) but it doesn't seem worth
it, because that would make the code more error-prone.