[LibOS] Use correct v5.8 x86 syscall numbers
When compiling against v5.8 (or later) kernel headers, we encounter
compilation errors such as:

  In file included from /usr/include/asm/unistd.h:13,
                   from shim_parser.c:11:
  shim_parser.c:408:10: error: array index in initializer exceeds array bounds
           [__NR_io_uring_setup]    = {.slow = 0, .parser = {NULL}},

The remedy is to extend `syscall_parser_table` by modyfing the
definition of LIBOS_SYSCALL_BOUND so the offending system calls fit in
the table.

While at it, also fix the actual values for syscall numbers in
"shim_syscalls.h" (the numbers are not sequential, there is an
intentional gap between the numers).