[LibOS] Fix incorrect use of LISTP macros in Async helper thread
The Async helper thread traverses through the `async_event` list of
scheduled events and copies the triggered events in the `triggered`
list. Theoretically, each event may be present in both lists at the
same time, so the event struct must have two `LIST_TYPE` fields.
However, there was a bug: both lists were described by a single
`list` field. In cases where the event's `list` was rewired to the
`triggered` list, it suddenly disappeared from the `async_event`
list, which led to a hang of Graphene.