[LibOS] Add a patch for a glibc building bug
Glibc does not properly escape path when passing it to sed in one of the
makefiles. Namely it uses @ as a separator in sed and if such character
happens to be part of the path (which sometimes happens on Jenkins), the
build blows with wonderful and totally uninformative errors.
This commit adds a Glibc patch to properly escape @ in paths.

Signed-off-by: Borys Popławski <borysp@invisiblethingslab.com>