[LibOS] Connect to IPC leader at process startup
Previously Graphene made IPC connections on demand and there was
a hidden assumption that the main process is an IPC leader that never
exits. Although there was some code to find new IPC leader in case the
old one dies, it was broken and only partialy used.
This commit causes Graphene processes to connect to the IPC leader on
startup, and disconnection of the leader is treated as a fatal error in
the child process.

Signed-off-by: Borys Popławski <borysp@invisiblethingslab.com>