[LibOS] Treat shim_handle.info as tagged union
This change establishes the convention that `hdl->info.xxx` can
only be accessed when `hdl->type == TYPE_XXX`. This is asserted
or checked whenever necessary.

To achieve that, we also remove TYPE_DIR handle type. Before,
the filesystem code modified `hdl->type` to TYPE_DIR when
encountering a directory, which clobbered existing information
about a handle. However, the data associated with old type
(e.g. `hdl->info.file`) was still being accessed.

This change converts TYPE_DIR to a boolean attribute in order to
preserve the original handle type. For filesystems using the
`pseudo_*` family of functions, a special TYPE_PSEUDO is
introduced for directories.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>