[Examples] python-simple: use `127.0.0.1` instead of `locahost` in test scripts
For some reason (needs more analysis and a proper fix), the `urllib3`
package used in Python 3.8 (which is the default on Ubuntu 20.04) and
backed by Glibc 2.33 tries to perform `socket(AF_INET6)` plus
`bind(<sock>, ::1:0)` (notice the zero port) plus `connect(<sock>)`.
This results in EADDRNOTAVAIL error when Graphene tries to perform
host-level `bind()`. This doesn't happen on Python 3.6 (default on
Ubuntu 18.04). This also doesn't happen on Glibc 2.31.

For now, we replace `localhost` (which needs to be resolved with the
help of Glibc's getaddrinfo) with IPv4 `127.0.0.1`.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>