[LibOS] Don't store dentry paths
This change removes `shim_dentry.rel_path` and `shim_handle.path`
fields. Instead, the path is computed whenever necessary. This also
means that the length limit for relative path is lifted.

To obtain a path for a dentry, use `dentry_abs_path` or
`dentry_rel_path`. The constructed path is allocated on heap and must be
freed afterwards.

Because computing the path can fail when out of memory, the functions
for hashing dentries do not compute the path explicitly, but follow the
dentry chain to the root.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>