[Pal/Linux-SGX] Rename "checksum" to "hash" and "stub" to "chunk"
Previously, the code interchangeably used "checksum" and "hash" when
referring to SHA-256 hashes, as well as "stub" and "chunk" when
referring to file chunks (16KB blocks of a file). This commit unifies
the naming to more correct "hash" and "chunk" everywhere. Note that the
`sgx.trusted_checksum` manifest syntax is kept for now because we don't
want to break old manifests.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>