[Common,Pal] Move `debug_map` from common lib to Linux-common PAL
The `debug_map.{c,h}` files maintain a debug map for GDB, and thus
conceptually do not belong to "common host-agnostic functionality".
Also, this functionality is only meaningful for PALs and should not be
exposed to LibOS. Finally, `debug_map.c` includes `spinlock.h` which in
turn includes LibOS-internal headers, thus previously leading to
compilation failures of the LibOS-agnostic common lib `graphene-lib.a`.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>