[LibOS] Do not update offset on chroot `FILE_DEV` files
There was a bug that LibOS manipulated the offset (`file->marker`) on
device files (with type `FILE_DEV`). This led to calls like
`DkStreamRead(.., /*offset=*/10, ..)` which ends up in PAL's device read
function that errors out because the offset is not zero (as expected).
Note that this special handling was already implemented for `FILE_TTY`
terminal devices, but we forgot about `FILE_DEV` generic devices.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>