[Pal,LibOS] Handle entrypoint in LibOS
This removes the leftover code for handling entrypoint from PAL. Now,
the `libos.entrypoint` manifest key is handled in LibOS, and there's no
need to pass the executable path to PAL when creating a new process.

Because the PAL regression tests run without LibOS, we keep the
`pal.entrypoint` manifest key to handle control directly to a test
binary.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>