[LibOS] Rework IDs management
This commit completely reworks the management of IDs, which are used as
process (PID) and thread (TID) IDs.
Now there are two mechanisms which govern IDs management:
- global map of ID ranges into processes currently owning them (only
  stored inside IPC leader),
- local structure tracking which IDs current process owns and which of
  them are already used.

IDs are given to a process in ranges/batches to improve thread creation
latency (and possibly fork - new process creation - although the latter
is usually slow anyway). IDs are given sequentially and not reused until
necessary (i.e. all other are/were already used). This is consistent
with Linux behavior and helps with potential PID reuse bugs in user
applications.

Signed-off-by: Borys Popławski <borysp@invisiblethingslab.com>