[LTP] Disable `recvmsg01` because of the non-fatal crash
Previously, `recvmsg01` was enabled with only a handful of subtests
required to pass. In this test, one of the subtests spawns a child
process that crashes loudly (because of the bug in Graphene). It is
better to completely disable this test than to have only "partially
working" test in our CI.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>