[LibOS] Fix setting of `O_NONBLOCK` flag at the write end of pipe
Currently in graphene we don't save the `O_NONBLOCK` flag for the
write end of the pipe. We set the `O_NONBLOCK` flag for the
PAL_HANDLE but this isn't propagated to the SHIM_HANDLE when
creating pipe. This commit fixes the corner case.

Also, updated the `pipe_nonblocking` regression test to catch this
corner case.

Signed-off-by: Vijay Dhanraj <vijay.dhanraj@intel.com>