[LibOS] Fix setting of `O_NONBLOCK` flag at the write end of pipe
Previously in Gramine we didn't save the `O_NONBLOCK` flag for the write
end of the pipe. We only set the `O_NONBLOCK` flag for the PAL handle
but this wasn't propagated to the LibOS (shim) handle when creating
pipe. This commit fixes the corner case. Also, `pipe_nonblocking`
regression test is updated to test this corner case.

Signed-off-by: Vijay Dhanraj <vijay.dhanraj@intel.com>