[Examples] Decouple all examples from the core Graphene repo
Examples are supposed to be stand-alone and work with packaged Graphene.
To this end, this commit removes the GRAPHENEDIR envvar and the
associated includes of Graphene-internal Makefiles. This commit also
copies the `download` bash script under `Examples/` so that it is closer
to the actual examples (this will be handy when moving examples out of
this core repository).

Ideally, `SGX_SIGNER_KEY` should be explicitly asked from the user, but
for now we default to Graphene-internal `Pal/src/host/Linux-SGX/signer`
for simplicity.

Also, `ra-tls-mbedtls` and `ra-tls-secret-prov` are still tightly
coupled to Graphene-internal paths. This will be fixed when Linux-SGX
PAL build is finally moved to Meson.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>