[Pal/Linux-SGX] Fix renaming issue with protected files
With certain workloads involving Federated Learning, the frameworks
create protected files and rename them as part of the flow. This does
not work currently with protected files because every protected file
maintains the path name in its metadata (stored when the file was first
created) and which is checked during open to make sure the incoming path
is the same as what is already stored in the metadata for this file.

A normal rename call does not need the file to be open in RW mode, but
since the metadata needs to be updated with the new path, we pass a
special PAL_OPTION_RENAME flag to DkStreamOpen during rename and open
the protected file in RW mode with required permissions, update the
metadata with the new path and flush it. Then the flow goes through the
one for normal file renaming. If the ocall to rename fails, we restore
the metadata back to what it was earlier so the file remains usable.

Signed-off-by: Sankaranarayanan Venkatasubramanian <sankaranarayanan.venkatasubramanian@intel.com>