[LibOS] Add UID and GID to manifest
This implementation enables Gramine to run an executable with another user than root. This becomes very handy in case of executables, which shouldn't run with the root user. The implementation of this feature sets also the effective uid/gid to the value, which is defined in the manifest for the directive loader.uid/gid