[LibOS] Get rid of `FDTYPE` and replace it with `uint32_t`
Previously, LibOS used `typedef uint16_t FDTYPE`. This was unnecessary
and hard-to-read. Moreover, it led to a bug in `__enlarge_handle_map()`:
the new size was of 64-bit type `size_t` but it was assigned to
`fd_size` field of 16-bit type `FDTYPE`, which led to silent truncation.
This resulted in failing asserts, because active FDs suddenly became
greater than the limit in the FD map.

This commit fixes this bug by changing the type from 16-bit to 32-bit,
and also removes the opaque and confusing `FDTYPE`.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>