[Pal/Linux] Memorize the assigned ephemeral port during `udp_bind()`
In Linux, if a port number (in the input UDP socket address) during
`bind()` is specified as 0, an ephemeral (random) port is assigned.
Previously, Linux PAL forgot to update the port in the UDP socket
address, so `getsockname()` always returned 0 instead of the actual
port. This commit fixes this bug.

Signed-off-by: Sonali Saha <sonali.saha@intel.com>