[Common,Pal,LibOS] Report crash location in debug builds
This change adds support for determining crash location (source line,
file, etc.) in situations where it cannot be known at compile time. This
is done by using the PAL debug maps to determine the right ELF file and
offset, and then running `addr2line` in a subprocess.

This feature is enabled in debug builds, and used in AddressSanitizer
and PAL/LibOS exception handlers.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>