[LibOS] Make `get_new_id` immediately move ID ownership if needed
Previously allocating ID for a new process did not immediately change
the ownership of this ID (it happened later in the child process). This
could cause issues when one thread does a `fork` syscall and another
exits, releasing its own thread ID - IPC leader might get notified to
release an ID range, which it thinks is a part of a bigger one, since it
was not yet notified about the ID ownership change.
This commit causes allocating a new ID to also immediately notify IPC
leader about ownership change, atomically w.r.t. other threads in the
current process.

Signed-off-by: Borys Popławski <borysp@invisiblethingslab.com>