[LibOS] Make sure LibOS doesn't have the executable stack flag
This ensures that the LibOS binary (`libsysdb.so`) is not incorrectly
marked as requiring executable stack.

Note that this doesn't have any visible effect right now: our ELF loader
does not respect the relevant program header (`GNU_STACK`), and under
SGX 1, we need to map all memory as RWX anyway.

Technical details: The linker clears the executable stack flag for a
binary only if none of the `.o` files require it. This is determined
automatically for C sources, but when compiling assembly, we need to
pass an extra `-Wa,--noexecstack` flag. This flag is already set
globally, but `vdso-data.o` is a Meson custom target, and doesn't use
global flags.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>