[Meson] Stop using `tomlc99_inc`
Using `tomlc99_inc` as an include directory is subtly broken: it
actually adds two directories (source and build), the first contains
original, unpatched `toml.h`, and the second contains the patched
version.

If a target does not have a dependency on the patched TOML files, it can
be compiled before the patched `toml.h` is built. As a result,
`#include "toml.h"` will use the unpatched file. Because the unpatched
file includes `<stdio.h>`, this can cause hard-to-debug build issues.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>