[Pal,LibOS] Use a pre-built symbol map for reporting location
After adding a seccomp filter for system calls, we no longer can execute
an external binary (`addr2line`). Instead, we dump a list of symbols to
a text file, and parse it when necessary.

Compared to the previous solution, we no longer report the source line
(only source file), and we only report detailed location for Gramine's
own binaries.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>