[Pal] Deprecate `loader.preload` and use `loader.entrypoint` instead
Previously, PAL had the ability to preload several libraries via
`loader.preload = "pal_lib1,pal_lib2,pal_lib3"` syntax. This ability was
supposed to be used as an interface for "PAL plugins". In practice, the
only loaded library (and simultaneously the entrypoint) was the LibOS.

This commit removes this unused ability. The syntax `loader.preload` is
deprecated and is currently used as an alias to `loader.entrypoint`. PAL
regression tests that tested this preload functionality are removed.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>