[Pal/Linux-SGX] Fix user address range calculation
Before, we reported the available range based on how much memory PAL has
already allocated. As a result, the range could be different for parent
and child processes, and sometimes that caused LibOS in the child
process to allocate memory in a range that would be needed by the forked
user process.

(This problem was not visible before the "Remove the 64MB static buffer"
commit, because by default PAL only used memory from the static buffer,
and did not deplete the available range.)

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>