[common,Pal,LibOS] ASan: Enable stack sanitization
The current implementation is able to catch stack buffer overflows and
use-after-scope bugs (but not use-after-return).

The implementation requires Gramine to maintain proper hygiene when
handling stack memory. Otherwise, false positives are possible: if
Gramine stops using a stack (e.g. when returning from a syscall), but
doesn't unpoison it, the poisoned stack memory can incorrectly trigger
ASan the next time that stack is used.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>