[Docs] Improve explanation for `sgx.enclave_size` manifest option
From the previous explanation, it looked like it is enough to enable
EDMM feature only in hardware, and the behavior of Gramine would change
automatically. In fact, the EDMM feature must be enabled in all
components of the hardware/software stack (including Gramine).

Also, it was confusing to readers to jump straight to RAM, without a
mention of SGX-specific EPC. This looked like EPC was some memory
separate from RAM.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>