[LibOS] Fix FUTEX_WAKE_OP's implementation of `ANDN` operation
Current code implements `ANDN` op as NAND operation but this op according
to man pages should do a `bitwise negation` of `oparg` and then `bitwise and`
with value of `uaddr2`. This commit fixes the issue.

This commit also updates the `futex_wake_op` regression test for
`FUTEX_OP_ANDN` and other operations as well.

Signed-off-by: Vijay Dhanraj <vijay.dhanraj@intel.com>