[Pal] Always map LibOS binary inside enclave memory
In the recent RTLD refactoring, the LibOS binary was initially mapped
into memory via `_DkStreamMap(..., /*prot=*/0, ...)`. This led to a
corner case in the Linux-SGX PAL: if `sgx.file_check_policy =
"allow_all_but_log"` was set in the manifest (or the LibOS binary was
marked in `sgx.allowed_files`), the LibOS binary was considered an
allowed file. And because protections were `0` (aka MAP_SHARED),
Linux-SGX's `file_map()` callback decided to map the file in untrusted
memory, which was incorrect and led to failures later.

This commit fixes this bug by specifying PAL_PROT_WRITECOPY -- this
guarantees that the file will be mapped inside enclave memory.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>