[tests] Work around LTP tests leaving running processes
The previous LTP runner only waited for the main process to exit. Right
now we wait for all processes, which uncovered a (known) problem with
signal delivery in Gramine: apparently LTP tests sometimes leave running
child processes.

Because these tests are still useful for us, we do not treat such
situations as a timeout, only log a warning.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>