[Pal/Linux-SGX] Allow non-existing files in trusted-files list
Previously, Gramine registered SGX trusted files during enclave
initialization and initialized the sizes of all trusted files via
`_DkStreamAttributesQuery()` which tries to access the file and fails if
the file doesn't exists. This commit defers this size initialization
until actual file open happens; this allows to skip non-existing files
(e.g., files that existed in a build container but were removed in the
slimmer version of this build container).

Signed-off-by: Jitender Kumar <jitender.kumar@intel.com>