[LibOS] Do not inherit an associated file during VMA split
Previously, a new VMA that was split from the old VMA would inherit the
file that was associated with the old VMA, even if the new VMA would
start past the end of the associated file. This led to obscure segfaults
during VMA checkpoint-and-restore on the Linux-SGX PAL. Also, it led to
worse performance because the checkpoint restore logic would try to
reopen and mmap the file associated with the new VMA.

This bug was found on Python's qiskit package (it led to a segfault).

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>