[Pal/Linux-SGX] Skip trusted-files verification on mmap past file end
There was a bug in Gramine-SGX implementation of trusted files:
invocation of `mmap(..., <trusted-file-fd>, <offset-past-file-end>)`
crashed Gramine process. Such an invocation can happen during checkpoint
restore in the child process. This commit fixes this bug by skipping
content verification on such mmap and simply returning success.

This bug was found on Python's qiskit package (it led to a segfault on
SGX PAL). This commit adds new LibOS regression test to catch this bug.
The test also checks protected-files logic: it was not affected by this
bug but we still add such check for future-proofing.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>