[LibOS] Skip mmapping past-file-end VMA contents in checkpoint restore
There was a bug in Gramine: if a VMA was file-backed and then split into
two VMAs, such that the second VMA starts past the end of the
corresponding file, this VMA info was sent to the child process (but not
the contents of the corresponding memory region). However, this VMA info
in the sent checkpoint had the `need_mapped` field that was interpreted
by the child process as "mmap from the file at given offset". This
operation is meaningless (because the given offset is past the end of
file) but allowed in Linux. However, for trusted/protected files on
Linux-SGX this leads to their SGX-specific checks failing, and Gramine
crashed. This commit fixes this bug.

This bug was found on Python's qiskit package (it led to a segfault).
This commit adds a new LibOS regression test to catch this bug.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>