[LibOS] Do not re-map private VMA contents in checkpoint restore
There was a bug in Gramine: tainted file-backed memory regions marked as
MAP_PRIVATE were not sent to the child process sometimes -- instead
these memory regions were re-mapped in the child from the corresponding
file. This commit forces Gramine LibOS to always send such memory
regions in the checkpoint and disallows them to be re-mapped by child.

This bug was found on Python's qiskit package (it led to a segfault on
SGX PAL). This commit adds new LibOS regression test to catch this bug.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>