[LibOS] Re-map tainted shared file-backed memory regions on fork
Previously, Gramine sent the contents of a tainted shared file-backed
memory region in a checkpoint blob (in a `shim_mem_entry` object). This
resulted in this memory region being allocated as anonymous via
`DkVirtualMemoryAlloc()` in the child process, so the updates on this
memory region were not reflected in parent/child.

Fixing this bug allows to remove the workaround for BROK LTP tests.
This is because LTP reports results of the child via the file `/dev/shm`
mapped as `MAP_SHARED`.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>