[Pal/Linux] Introduce `sys.enable_host_vdso` manifest option
There is a bug in the Linux kernel: the placement of the vDSO library is
randomized even when app is started with `ADDR_NO_RANDOMIZE`. This bug
manifests itself on systems with 5-level paging enabled (e.g., newer
Icelake Intel CPUs).

The only way to completely circumvent this bug is to not use the vDSO
library. This commit changes default Gramine behavior to disable host
vDSO and introduces `sys.enable_host_vdso` to re-enable it. This
manifest option does not apply to the SGX environment (because it
doesn't use host vDSO anyway).

For the initial attempt to fix this bug, see commit e7429e13bb3e9
"[Pal/Linux] Add a workaround for a kernel bug randomizing VDSO".
Unfortunately, that bug fix does not eliminate the bug completely but
only makes it less frequent.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>