[common] Fix mbedtls configuration mismatch
This ensures that userspace programs using the mbedtls library (e.g.
`gramine-sgx-pf-crypt`) use the same configuration flags as the the
mbedtls library itself.

The problem was that the Meson build applied a patch to a copy of
`mbedtls/config.h`, but the programs included the unpatched version. As
a result, the both sides disagreed about size of some data structures,
causing `gramine-sgx-pf-crypt` to crash (when compiled with Clang in
release mode).

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>