[LibOS] Convert `fs.mount` to an array
This converts the `fs.mount` manifest key to an array instead of table.
Similar to `sgx.{trusted,allowed,protected}_files`, an array can be
parsed more efficiently. It's also more natural, because order matters
when specifying mount points.

The old (table) syntax is still accepted, but triggers a deprecation
warning.

This commit converts the main manifest for regression tests. Other
manifests will be converted later.

In addition, Gramine now allows omitting the `type` key (with "chroot"
as default) and the `uri` key (so that it's easier to use the `tmpfs`
filesystem).

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>