[LibOS,Pal/Linux-SGX] Don't use RBX as operand in inline asm
It seems that in LLVM, the RBX register is reserved, and using it as
inline assembly operand can produce unexpected results. In our case,
Clang emitted incorrect code around the ENCLU instruction when compiling
with ASAN and optimizations enabled.

The ENCLU instructions are now defined in a separate assembly file. In
addition, RBX use is removed from one inline assembly fragment in LibOS.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>