[LibOS,Pal/Linux-SGX] Don't use RBX as operand in inline asm
It seems that in LLVM, the RBX register is reserved, and using it as
inline assembly operand can produce unexpected results. In our case,
Clang emitted incorrect code around the ENCLU instruction when compiling
with ASAN and optimizations enabled.

To fix the problem, I'm moving the ENCLU instructions to separate
functions defined in assembly files. I'm also updating one instance of
RBX use in LibOS.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>