[LibOS] Propagate errors when writing to pseudo-files
Before, the `save` callback on pseudo-files was invoked on flush and
close. However, this prevented errors from being reported to the user,
since Gramine's `close` syscall does not actually report any errors.

As a result, the user could write a garbage data to `/dev/attestation`
files, and `write` and `close` syscalls would both succeed.

We could return an error from the `close` syscall, but this is not a
good solution: the user might close a duplicated file descriptor (which
actually does not close the handle), or even exit without closing the
handle. Both of these circumstances happen in shells (e.g. Bash) when
executing `echo text > file`.

Instead, we do what Linux does, and trigger a callback on every `write`
operation, not on file flush/close.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>