Rewrite sysfs topology support
This commit completely rewrites all three parts of sysfs topology handling:
 - parsing sysfs files on the host,
 - passing the parsed topology information to LibOS, possibly across a
   security boundary,
 - generating sysfs files inside LibOS based on the topology
   information.

The overall idea of the rewrite is to simplify and normalize the data
structures passing through the untrusted interface, to ensure they are
properly handled and sanitized. Effectively, this commit shifts as much
complexity as possible to either untrusted side or to the part working
on already sanitized, in-enclave data.

The reason for this change is the number of security issues found in the
old code and review difficulty of sanitization in the old approach.

Additionally, this commit adds support for offline CPU threads (but no
hot-plug, the state is snapshotted at the start).

Signed-off-by: Michał Kowalczyk <mkow@invisiblethingslab.com>