[Pal/Linux-SGX] Allow READ+WRITE mmap of protected files selectively
Mmap() on protected files only allows PROT_READ or PROT_WRITE, not both.
However, when MAP_PRIVATE flag is present, changes to the file do not need
to be written back to the file, essentially makes the behavior of file
access identical to the read only mmap.

By allowing READ+WRITE mmap with MAP_PRIVATE flag, it is now possible to
load executables from the protected FS.

Fixes #196

Signed-off-by: Lejun Zhu <lejun.zhu@gmail.com>