[tests] Respect the change dir option
The `-C` option (change dir) and `-n` option (configure file name)
are incompatible. The gramine-test first confirms the file's
existence and later changes the current working directory.

The verification is done by option library predefined
type `click.Path`. The verification is done before the gramine-test
changes the directory, so it always checks the file's
existence in the current working directory.

The workaround for this problem is to use a callback.
When the `-C` option is passed, change the directory as fast as
possible and not wait until the command function is called.
Additionally, the argument `is_eager` will cause the library to
parse the change dir option before anything else.

Signed-off-by: Mariusz Zaborski <oshogbo@invisiblethingslab.com>