[Pal/Linux-SGX] tools: Fix resource leak in Secret Prov lib
The resources allocated for mbedTLS functioning on the client side of
Secret Prov lib (the one which attests itself to the remote party and
runs inside the Gramine-SGX enclave) were not freed if the user of the
Secret Prov lib used `secret_provision_start(..., &ctx)` API. This led
to memory and resource leaks. In particular, the TLS connection to the
remote party was shutdown but the underlying FD was not closed.

Kudos to Liang Fang <liang.a.fang@intel.com> who reported this leak.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>