[Pal/Linux-SGX] Add MAA to RA-TLS and SecretProv libs
Microsoft Azure Attestation (MAA) is a new attestation scheme available
in the MS Azure public cloud. It uses the classic DCAP SGX quotes but
replaces the DCAP/EPID quote-verification software infrastructure with
the new MAA attestation provider infrastructure. In particular, the
remote user is supposed to send the SGX quote in the HTTPS REST
attestation request to the MAA attestation provider and receive back the
JSON Web Token (JWT) that contains claims about the attesting SGX
enclave.

This commit introduces new `_maa` versions of the RA-TLS and Secret
Provisioning verification libraries. The documentation and the
`ra-tls-mbedtls` and `ra-tls-secret-prov` examples are updated.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>