Use `#pragma once` instead of handcrafted header guards
Cons of `#pragma once`:
    - not in C standard
    - compiler-specific behavior in cases like including the same file
      through different {soft,hard}links, or multiple copies of the
      same header in the repo
Pros:
    - less boilerplate
    - works well in practice
    - no need to update it when moving/renaming a header file
    - handcrafted header guards are error prone (easy to have
      non-matching `#ifndef` and `#define`, easy to collide with other
      file header guards, and especially prone to copy-paste errors)

The cases with {soft,hard}links and using multiple copies of the same
headers are rather exotic and we don't ever plan to have them in our
codebase.

Overall: `#pragma once` works well in practice and it's much more likely
that we'll introduce a bug with handcrafted include guards than with
`#pragma once`.

Signed-off-by: Michał Kowalczyk <mkow@invisiblethingslab.com>