[LibOS,PAL] Add flexible device-specific IOCTL support
The newly added `ioctl()` syscall emulation on device-backed file
descriptors is pass-through. It is insecure by itself since the
emulation only passes the arguments to and from the untrusted memory:
- IOCTL arguments are passed as-is from the app to the untrusted host,
  which may lead to leaks of secret data;
- untrusted host can change IOCTL arguments as it wishes when passing
  them from Gramine to the device and back.

It is the responsibility of the app developer to correctly use IOCTLs,
with security implications in mind. In most cases, IOCTL arguments should
be encrypted or integrity-protected with a key pre-shared between
Gramine and the device.

On the Linux-SGX PAL, a set of IOCTL requests must be explicitly allowed
in the manifest via the new option `sgx.allowed_ioctls`.
Also, the allowed IOCTLs' arguments (typically pointers to complex
nested objects) must be explicitly described in the manifest via the new
option `sgx.ioctl_structs.[identifier]` and a corresponding reference in
`sgx.allowed_ioctls`; see docs for explanation of the IOCTL struct
format.

This commit adds a new LibOS test `device_ioctl` that tests the flexible
IOCTL logic against Gramine dummy device `/dev/gramine_test_dev`. This
device is found in companion repo `gramineproject/device-testing-tools`.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>