[CI] Install missing `bc` tool
Additionally, observed that `lighttpd` workload in our CI-Examples
calculates the max_conn to be `274` but we set up the number of
connections to be `300` resulting in the following log message.

```
server.c.980) [note] sockets disabled, connection limit reached
server.c.973) [note] sockets enabled again
```

To address this resource limitation, set max-connections to `1024`
in the `lighttpd-generic.conf` as part of this commit.

Signed-off-by: Vijay Dhanraj <vijay.dhanraj@intel.com>