[CI-Examples] Modify secret_prov_server to not require a key
In our ra-tls-secret-prov example, we have one server (verifier of the
SGX enclave and releaser of the secrets) and three different kinds of
clients (three SGX enclaves, each showcases a different possible flow).
But only one of the clients actually needs a real Encrypted Files key
to be provisioned (this client is secret_prov_pf_client).

Signed-off-by: aneessahib <anees.a.sahib@intel.com>