[common] Don't extend files while reading data
The file should grow only during write and truncation.
The protected file was extended when the application opened it for
reading and writing and read from an offset larger than the current size.
Previous logic, in the `ipf_update_after_seek` function, assumed that
if the file is writable, the file should be extended.
This function was called in the read and write routine of protected files.

New tests verify that the file is not extended after seek/tell routine,
and they are not extended after the read syscall.

Signed-off-by: Mariusz Zaborski <oshogbo@invisiblethingslab.com>