[python] Use elftools to parse ELF
Remove parsing ELF with readelf, instead use pyelftools, which are
already installed for GDB integration.

The dependency on python3-pyelftools is not strictly necessary in all
cases, technically it's not needed for gramine-direct without running
gdb (and tests which also run gdb). However, it's probably easier to
write and read docs this way.

The immediate reason for this change was that binutils (readelf)
dependency was missing in the package, but lets fix it the proper way.

Signed-off-by: Wojtek Porczyk <woju@invisiblethingslab.com>