[Docs] Change "GSC" page to "Container integration"
We now have not only GSC, but also the Gramine Docker image hosted on
DockerHub. Update our documentation to mention both tools and prepare
for more tools/workflows in the future. Also, GSC was split from the
core repository a long time ago and there is no need to emphasize that,
so we rephrase its description.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>