[LibOS] Override `argv[0]` with file pathname for interpreter scripts
According to POSIX, if the pathname argument of `execve()` specifies
an interpreter script, then interpreter will be invoked with the following
arguments:

    interpreter [optional-arg] pathname arg...

where `pathname` is the pathname of the file specified as the first argument of
`execve()`, and `arg...` is the series of words pointed to by the `argv` argument
of `execve()`, starting at `argv[1]`.

Signed-off-by: aneessahib <anees.a.sahib@intel.com>

fixup! Override argv[0] with pathname in execve

Signed-off-by: aneessahib <anees.a.sahib@intel.com>

[LibOS] Override `argv[0]` with file pathname for interpreter scripts

Signed-off-by: aneessahib <anees.a.sahib@intel.com>

fixup! [LibOS] Override `argv[0]` with file pathname for interpreter scripts

Signed-off-by: aneessahib <anees.a.sahib@intel.com>

fixup! [LibOS] Override `argv[0]` with file pathname for interpreter scripts

Signed-off-by: aneessahib <anees.a.sahib@intel.com>

fixup! fixup! [LibOS] Override `argv[0]` with file pathname for interpreter scripts

Signed-off-by: aneessahib <anees.a.sahib@intel.com>