[Pal] Always allocate `file.realpath` and `dir.realpath` in separate memory locations
Previously in `file_open`, `file.realpath` is allocated together with the handle. But in `file_rename`, it is allocated to a separate memory location. After this change, `file.realpath` is also allocated separately in `file_open`, and there is no need to check the address of the pointer when doing free inside `file_close`. We also set the pointer to `NULL` after free to avoid double free. Same for `dir.realpath`.

In addition, we copy `dir.realpath` before doing `dir_close` inside `dir_delete` to avoid using after free of this pointer.

Signed-off-by: Yao Ji <yao.ji@intel.com>