[tools/sgx,curl,Docs] Minimize and build libcurl statically to link with `tools/sgx`
Previously, we built `libsgx_util` as a shared library which is linked
into RA-TLS, Secret Provisioning libraries and some other tools. This
library relies on libcurl, taken from the host platform. However,
libcurl has a high number of dependencies itself, most of which are not
required by Gramine, only consume space and are hard to track and reason
about. This is especially pronounced for the RA-TLS/SecretProv libs
because they typically run inside SGX enclaves and therefore must be
minimal and reproducible.

This patch:
- builds `libsgx_util` into a static library
- builds and minimizes `curl` into a static library (leveraging
  `mbedtls` as the TLS backend)
- links all dependencies of `tools/sgx` statically to avoid pulling in a
  lot of dependencies.

Signed-off-by: Kailun Qin <kailun.qin@intel.com>