[LibOS] Use proper const qualifiers for `argv` and `envp`
Previously, LibOS codebase used incorrect types:
- `const char** argv`
- `const char** envp`

In reality, the types must be "arrays of constant pointers". So the
correct types should be:
- `char* const* argv` (array of constant ptrs to data)
- `const char* const* envp` (array of constant ptrs to constant data)

Notice that argv items point to non-constant data: this is because argv
list is modified in LibOS -- both the pointers and the strings. On the
other hand, envvars are not modified so they have const qualifiers in
both places.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>