[LibOS,Pal] Clean-up sched set/get affinity functions
Current implementation of sched set/get affinity has a wierd
syntax of passing cpumask array as unsigned long but number of
elements in the array as bytes. This is misleading. This commit
fixes the issue. This commit also reorders the cpumask and
cpumask_size parameters  in the PAL set/get affinity functions.

Signed-off-by: Vijay Dhanraj <vijay.dhanraj@intel.com>