[tools/sgx] Initialize `strtok_r()`'s saveptr to NULL
From the man page of `strtok_r()`
(https://man7.org/linux/man-pages/man3/strtok.3.html#NOTES):
on some implementations, `*saveptr` is required to be NULL on the first
call to `strtok_r()` that is being used to parse str.

This patch thus initializes the saveptr of `strtok_r()` to NULL for
better portability.

Signed-off-by: Kailun Qin <kailun.qin@intel.com>