[tools/sgx] Fix `pf_crypt` tool to be runnable within an enclave
The current implementation of the `pf_crypt` tool uses `fstat`
to find `wrap_key` length. But `fstat` implementation in Gramine
doesn't report the size field for "/dev" pseudo-files. So this
commit replaces the use of `fstat` with `lseek` to determine the
length of the file to make it run in Gramine.

Signed-off-by: Vijay Dhanraj <vijay.dhanraj@intel.com>