[tools] Terminate whole process upon Secret Provisioning failure
If Secret Provisioning is called via a constructor (explicit
`loader.env.SECRET_PROVISION_CONSTRUCTOR = "true"` in the manifest), the
app is assumed to be completely unmodified. Thus, there is no graceful
way to terminate the process, so the constructor itself has to kill the
process with an error message when Secret Provisioning fails.

Signed-off-by: aneessahib <anees.a.sahib@intel.com>